+++ lib/poodr_rspec/bicycle.rb
@@ -4,8 +4,8 @@ module PoodrRspec
 
     def initialize(args = {})
       @size = args[:size]
-      @chain = args[:chain]
-      @tire_size = args[:tire_size]
+      @chain = args[:chain] || default_chain
+      @tire_size = args[:tire_size] || default_tire_size
       post_initialize(args)
     end